+++ UniverseEngine/UniverseEngine/Rect2D.h
@@ -5,10 +5,10 @@
 namespace UniverseEngine {
     struct Rect2D {
         Rect2D() = default;
-        Rect2D(uint32_t width, uint32_t height) : offset{}, extent(width, height) {
+        Rect2D(float width, float height) : offset{}, extent(width, height) {
         }
 
-        glm::uvec2 offset;
-        glm::uvec2 extent;
+        glm::vec2 offset;
+        glm::vec2 extent;
     };
 }  // namespace UniverseEngine
